src/db.py
import os
from motor.motor_asyncio import AsyncIOMotorClient
from dotenv import load_dotenv

load_dotenv()

MONGODB_URI = os.getenv("MONGODB_URI", "mongodb://localhost:27017")
MONGODB_DB = os.getenv("MONGODB_DB", "apiproyect")

client = AsyncIOMotorClient(MONGODB_URI)

db = client[MONGODB_DB]


async def get_collection(name: str):
    return db[name]
