requirements.txt
fastapi==0.111.2
uvicorn[standard]==0.23.2
motor==4.4.0
pydantic==2.8.0
python-dotenv==1.0.1
pytest==8.4.3
httpx==0.27.0
